+++ outsystems/pipeline/fetch_lifetime_data.py
@@ -1,5 +1,7 @@
 # Python Modules
-import os, sys, argparse
+import os
+import sys
+import argparse
 
 # Workaround for Jenkins:
 # Set the path to include the outsystems module
@@ -16,7 +18,8 @@ from outsystems.lifetime.lifetime_base import build_lt_endpoint
 from outsystems.vars.lifetime_vars import LIFETIME_HTTP_PROTO, LIFETIME_API_ENDPOINT, LIFETIME_API_VERSION
 from outsystems.vars.file_vars import ARTIFACT_FOLDER
 
-############################################################## SCRIPT ##############################################################
+
+# ---------------------- SCRIPT ----------------------
 def main(artifact_dir: str, lt_http_proto: str, lt_url: str, lt_api_endpoint: str, lt_api_version: int, lt_token: str):
     # Builds the LifeTime endpoint
     lt_endpoint = build_lt_endpoint(
@@ -24,10 +27,10 @@ def main(artifact_dir: str, lt_http_proto: str, lt_url: str, lt_api_endpoint: st
 
     # Get Environments
     get_environments(artifact_dir, lt_endpoint, lt_token)
-    print("OS Environments data retrieved successfully.")
+    print("OS Environments data retrieved successfully.", flush=True)
     # Get Applications without extra data
     get_applications(artifact_dir, lt_endpoint, lt_token, False)
-    print("OS Applications data retrieved successfully.")
+    print("OS Applications data retrieved successfully.", flush=True)
 
 
 if __name__ == "__main__":