+++ outsystems/pipeline/tag_apps_based_on_manifest_data.py
@@ -1,0 +1,92 @@
+# Python Modules
+import sys
+import os
+import argparse
+
+# Workaround for Jenkins:
+# Set the path to include the outsystems module
+# Jenkins exposes the workspace directory through env.
+if "WORKSPACE" in os.environ:
+    sys.path.append(os.environ['WORKSPACE'])
+else:  # Else just add the project dir
+    sys.path.append(os.getcwd())
+
+# Custom Modules
+# Variables
+from outsystems.vars.file_vars import ARTIFACT_FOLDER
+from outsystems.vars.lifetime_vars import LIFETIME_HTTP_PROTO, LIFETIME_API_ENDPOINT, LIFETIME_API_VERSION
+
+# Functions
+from outsystems.file_helpers.file import load_data
+from outsystems.lifetime.lifetime_environments import get_environment_key
+from outsystems.lifetime.lifetime_base import build_lt_endpoint
+from outsystems.lifetime.lifetime_applications import set_application_version
+
+
+# ############################################################# SCRIPT ##############################################################
+
+def main(artifact_dir: str, lt_http_proto: str, lt_url: str, lt_api_endpoint: str, lt_api_version: int, lt_token: str, dest_env: str, app_list: list, dep_manifest: list):
+    # Builds the LifeTime endpoint
+    lt_endpoint = build_lt_endpoint(lt_http_proto, lt_url, lt_api_endpoint, lt_api_version)
+    # Get the environment keys
+    dest_env_key = get_environment_key(artifact_dir, lt_endpoint, lt_token, dest_env)
+
+    for deployed_app in dep_manifest:
+        if deployed_app["ApplicationName"] in app_list:
+            set_application_version(lt_endpoint, lt_token, dest_env_key, deployed_app["ApplicationKey"], deployed_app["ChangeLog"], deployed_app["Version"])
+            print("{} application successuflly tagged as {} on {}".format(deployed_app["ApplicationName"], deployed_app["Version"], dest_env), flush=True)
+
+# End of main()
+
+
+if __name__ == "__main__":
+    # Argument menu / parsing
+    parser = argparse.ArgumentParser()
+    parser.add_argument("-a", "--artifacts", type=str, default=ARTIFACT_FOLDER,
+                        help="Name of the artifacts folder. Default: \"Artifacts\"")
+    parser.add_argument("-u", "--lt_url", type=str, required=True,
+                        help="URL for LifeTime environment, without the API endpoint. Example: \"https://<lifetime_host>\"")
+    parser.add_argument("-t", "--lt_token", type=str, required=True,
+                        help="Token for LifeTime API calls.")
+    parser.add_argument("-v", "--lt_api_version", type=int, default=LIFETIME_API_VERSION,
+                        help="LifeTime API version number. If version <= 10, use 1, if version >= 11, use 2. Default: 2")
+    parser.add_argument("-e", "--lt_endpoint", type=str, default=LIFETIME_API_ENDPOINT,
+                        help="(optional) Used to set the API endpoint for LifeTime, without the version. Default: \"lifetimeapi/rest\"")
+    parser.add_argument("-d", "--destination_env", type=str, required=True,
+                        help="Name, as displayed in LifeTime, of the destination environment where you want to deploy the apps. (if in Airgap mode should be the hostname of the destination environment where you want to deploy the apps)")
+    parser.add_argument("-l", "--app_list", type=str, required=True,
+                        help="Comma separated list of apps you want to deploy. Example: \"App1,App2 With Spaces,App3_With_Underscores\"")
+    parser.add_argument("-f", "--manifest_file", type=str, required=True,
+                        help="Manifest file path, used if you have a split pipeline for CI and CD, where the CI pipeline will generate the deployment manifest file.")
+
+    args = parser.parse_args()
+
+    # Parse the artifact directory
+    artifact_dir = args.artifacts
+    # Parse the API endpoint
+    lt_api_endpoint = args.lt_endpoint
+    # Parse the LT Url and split the LT hostname from the HTTP protocol
+    # Assumes the default HTTP protocol = https
+    lt_http_proto = LIFETIME_HTTP_PROTO
+    lt_url = args.lt_url
+    if lt_url.startswith("http://"):
+        lt_http_proto = "http"
+        lt_url = lt_url.replace("http://", "")
+    else:
+        lt_url = lt_url.replace("https://", "")
+    if lt_url.endswith("/"):
+        lt_url = lt_url[:-1]
+    # Parte LT API Version
+    lt_version = args.lt_api_version
+    # Parse the LT Token
+    lt_token = args.lt_token
+    # Parse Destination Environment
+    dest_env = args.destination_env
+    # Parse App list
+    _apps = args.app_list
+    apps = _apps.split(',')
+    # Parse Manifest file
+    manifest_file = load_data("", args.manifest_file)
+
+    # Calls the main script
+    main(artifact_dir, lt_http_proto, lt_url, lt_api_endpoint, lt_version, lt_token, dest_env, apps, manifest_file)